Create README.md
An HR attendance dashboard is a visual representation of key employee attendance metrics. It can be used to track trends, identify areas for improvement, and make informed decisions about HR policies and procedures.
HR attendance dashboards typically include metrics such as:
🔹 Overall attendance percentage
🔹Employee absence rates
🔹Sick leave usage(SL)
🔹WFH trends(WORK FROM HOME).
📈 Insights:
✔ Presence trend: Employees tend to be more present in the office on Tuesdays and Thursdays, with the highest presence rates on June 7, 2022 (97.44%) and May 11, 2022 (97.40%), respectively. The lowest presence rate was on May 13, 2022 (77.92%).

✔ WFH trend: The highest WFH rates were on Friday, May 6, 2022 (23.44%) and June 17, 2022 (16.31%). The lowest WFH rate was on June 6, 2022 (3.95%).

✔ Sick leave trend: The highest sick leave rates were on Monday, May 30, 2022, Monday, June 13, 2022, and Tuesday, May 10, 2022 (all 5.42%). The lowest sick leave rate was on Tuesday, May 10, 2022 (3.25%).

📉 Recommendations:

🔼 Presence trend:
Consider offering more flexible work arrangements on Mondays and Fridays, when presence rates are typically lower. This could include allowing employees to start and end their workdays earlier or later, or to work from home one or two days per week.
Identify and address any factors that may be contributing to low presence rates on certain days of the week. For example, if employees are more likely to be late on Mondays because they have difficulty getting to work on time, consider offering a later start time on Mondays.

🔼WFH trend:
Consider offering a more permanent WFH policy, as employees appear to be taking advantage of the opportunity to work from home when it is available.
Provide employees with the resources and support they need to be successful when working from home. This could include providing laptops, headsets, and other necessary equipment, as well as training on how to use video conferencing and other communication tools.

🔼Sick leave trend:
Monitor sick leave rates for trends and identify any patterns. For example, if sick leave rates are consistently higher on certain days of the week or during certain times of the year, investigate the underlying causes and take steps to address them.
Promote a healthy work-life balance and encourage employees to take time off when they are sick. This could involve offering unlimited sick leave or providing paid time off for mental health days.

🔹Overall, HR attendance dashboards are a valuable tool for organizations of all sizes. They can help to improve employee attendance, productivity, and morale.
below is my portfolio link and also to have access to the dataset.